refactor(web/auth): introduce LoginButton component and remove LoginLogoutButton
- Added a new LoginButton component for streamlined user login functionality.
- Removed the obsolete LoginLogoutButton component to simplify the authentication interface in the navbar.
- Updated Navbar02 to utilize the new LoginButton, enhancing code clarity and maintainability.
