+++ apps/web/components/navbar-02/navbar-02.tsx
@@ -3,7 +3,7 @@
 import { NavMenu } from "./nav-menu";
 import { NavigationSheet } from "./navigation-sheet";
 import { ThemeSwitcher } from "@/components/theme-switcher";
-import { LoginLogoutButton } from "@/components/auth/login-logout-button";
+import { LoginButton } from "@/components/auth/login-button";
 import { PicoLogo } from "@/components/pico-logo";
 
 const Navbar02Page = () => {
@@ -19,7 +19,7 @@ const Navbar02Page = () => {
 
         <div className="flex items-center gap-3">
           <ThemeSwitcher />
-          <LoginLogoutButton />
+          <LoginButton />
 
           {/* Mobile Menu */}
           <div className="md:hidden">
